requirements.txt
numpy>=1.21.0
pandas>=1.3.0
MetaTrader5>=5.0.0
scikit-learn>=1.0.0
plotly>=5.0.0
pytest>=6.0.0
black>=21.0
flake8>=3.9.0
mypy>=0.910
python-dotenv>=0.19.0
ccxt>=1.50.0
ta>=0.7.0
websockets>=10.0
redis>=3.5.3
psycopg2-binary>=2.9.0
pika>=1.2.0
requests>=2.26.0
textblob>=0.15.3
newsapi-python>=0.2.6
beautifulsoup4>=4.9.3
nltk>=3.6.3
tensorflow>=2.8.0
torch>=1.9.0
stable-baselines3>=1.5.0
gymnasium>=0.26.0
optuna>=2.10.0
scipy>=1.7.0
seaborn>=0.11.0
matplotlib>=3.4.0
backtrader>=1.9.76
aiohttp>=3.8.0
transformers>=4.15.0
alpha_vantage>=2.3.1
finbert-sentiment>=0.1.0
